+++ master_server/vrp_log.h
@@ -1,5 +1,5 @@
 /*
-	VarastoRobo master server version 1.0.0 2019-12-10 by Santtu Nyman.
+	VarastoRobo master server version 1.1.0 2019-12-12 by Santtu Nyman.
 	github repository https://github.com/Jarno-Poikonen/VarastoRobo
 */
 
@@ -10,24 +10,11 @@
 extern "C" {
 #endif
 
+#include "vrp_master_server_types.h"
 #include <stddef.h>
 #include <stdint.h>
 #include <errno.h>
 #include <Windows.h>
-
-typedef struct vrp_log_t
-{
-	HANDLE handle;
-	uint32_t line_count;
-	size_t line_table_memory_granularity;
-	size_t line_table_memory_reserved;
-	size_t line_table_memory_commited;
-	struct
-	{
-		uint32_t offset;
-		uint32_t size;
-	}* line_table;
-} vrp_log_t;
 
 void vrp_close_log(vrp_log_t* log);
